HW-8, refactoring, moved app files to the main directory
