+++ src/main.jsx
@@ -1,5 +1,5 @@
 import { createRoot } from 'react-dom/client';
-import { App } from './components/app';
+import { App } from './app';
 
 const root = document.getElementById('root');
 const reactRoot = createRoot(root);
